don't return a local variable from get_from_cache
get_from_cache was reading cache data into a local char[] allocated
on the stack, and was returning the address of this local variable
to the caller. strdup the string before returning it. However this
leaks memory since I didn't review the call sites (but the program
is short lived so it should be acceptable).